Fix profile user

+++ app/Console/Commands/DeleteGameUnused.php
@@ -5,10 +5,10 @@ namespace App\Console\Commands;
 use App\Repositories\GameRepository;
 use Illuminate\Console\Command;
 
-class DeleteImageUnused extends Command
+class DeleteGameUnused extends Command
 {
-    protected $signature = 'app:delete-image-unused';
-    protected $description = 'Delete image of game unused with status 0';
+    protected $signature = 'app:delete-game-unused';
+    protected $description = 'Delete all of game unused with status 0 and empty thumbs';
     protected $gameRepository;
 
     public function __construct(GameRepository $gameRepository)
@@ -32,23 +32,38 @@ class DeleteImageUnused extends Command
                     continue;
                 }
 
+                $parse = parse_url($game['link']);
+                if (!array_key_exists('host', $parse)) {
+                    if (!empty($game['thumbs'])) {
+                        continue;
+                    }
+                }
+
                 if (!empty($game['icon'])) {
                     $pathIcon = public_path() . $game['icon'];
-                    unlink($pathIcon);
+                    if (file_exists($pathIcon)) {
+                        unlink($pathIcon);
+                    }
                 }
 
                 if (!empty($game['background'])) {
                     $pathBackground = public_path() . $game['background'];
-                    unlink($pathBackground);
+                    if (file_exists($pathBackground)) {
+                        unlink($pathBackground);
+                    }
                 }
 
                 if (!empty($game['thumbs'])) {
                     $pathThumb = public_path() . $game['thumbs'];
-                    unlink($pathThumb);
+                    if (file_exists($pathThumb)) {
+                        unlink($pathThumb);
+                    }
                 }
+
+                $game->delete();
             }
 
-            dump("---------Deleted all images of game on sever---------");
+            dump("---------Deleted all game unused on sever---------");
         } catch (Exception $e) {
             dump("---------" . $e . "---------");
         }
